requirements.txt
# Core AWS dependencies
boto3>=1.34.0
botocore>=1.34.0

# Database connectivity
pymysql>=1.1.0
sqlalchemy>=2.0.0

# Data processing
pandas>=2.0.0

# Configuration and utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Timezone handling
tzlocal>=5.0

# HTTP requests (if needed)
requests>=2.31.0

